refactor some more things
