+++ src/components/WordGrid/WordLineComponent.vue
@@ -24,4 +24,4 @@ const filled = computed(() => {
 });
 </script>
 
-<style lang="scss"></style>
+<style lang="scss" scoped></style>
